clean up migrations

+++ db/migrate/20220520165533_create_orders.rb
@@ -5,7 +5,7 @@ class CreateOrders < ActiveRecord::Migration[7.0]
       t.string :user_id, null: false
       t.integer :rating
       t.text :delivery_details
-      t.integer :total_price, default: 0, null: false
+      t.bigint :total_price, default: 0, null: false
       t.integer :order_items_quantity, default: 0, null: false
 
       t.timestamps
