+++ db/migrate/20220520165550_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration[7.0]
   def change
     create_table :products do |t|
       t.string :name
-      t.integer :price
+      t.bigint :price
       t.string :image_url
       t.string :category, null: false, default: 'food'
       t.text :description
